Added support for Ubuntu 16.04 and fixed a number of bugs

+++ methods/in_osx.rb
@@ -49,28 +49,30 @@ end
 # Check PF is configure on OS X 10.10
 
 def check_osx_pfctl(gw_if_name,if_name)
-  check_osx_ip_forwarding(gw_if_name)
-  pf_file = $work_dir+"/pfctl_config"
-  if File.exist?(pf_file)
-    File.delete(pf_file)
+  output = check_osx_ip_forwarding(gw_if_name)
+  if !output.match(/pass quick on #{gw_if_name}|any to any via #{gw_if_name}/)
+    pf_file = $work_dir+"/pfctl_config"
+    if File.exist?(pf_file)
+      File.delete(pf_file)
+    end
+    output = File.open(pf_file,"w")
+    if $verbose_mode == 1
+      puts "Information:\tEnabling forwarding between "+gw_if_name+" and "+if_name
+    end
+    output.write("nat on #{gw_if_name} from #{if_name}:network to any -> (#{gw_if_name})\n")
+    output.write("pass inet proto icmp all\n")
+    output.write("pass in on #{if_name} proto udp from any to any port domain keep state\n")
+    output.write("pass in on #{if_name} proto tcp from any to any port domain keep state\n")
+    output.write("pass quick on #{gw_if_name} proto udp from any to any port domain keep state\n")
+    output.write("pass quick on #{gw_if_name} proto tcp from any to any port domain keep state\n")
+    output.close
+    message = "Enabling:\tPacket filtering"
+    command = "sudo sh -c \"pfctl -e\""
+    execute_command(message,command)
+    message = "Loading:\yFilters from "+pf_file
+    command = "sudo sh -c \"pfctl -F all -f #{pf_file}\""
+    execute_command(message,command)
   end
-  output = File.open(pf_file,"w")
-  if $verbose_mode == 1
-    puts "Information:\tEnabling forwarding between "+gw_if_name+" and "+if_name
-  end
-  output.write("nat on #{gw_if_name} from #{if_name}:network to any -> (#{gw_if_name})\n")
-  output.write("pass inet proto icmp all\n")
-  output.write("pass in on #{if_name} proto udp from any to any port domain keep state\n")
-  output.write("pass in on #{if_name} proto tcp from any to any port domain keep state\n")
-  output.write("pass quick on #{gw_if_name} proto udp from any to any port domain keep state\n")
-  output.write("pass quick on #{gw_if_name} proto tcp from any to any port domain keep state\n")
-  output.close
-  message = "Enabling:\tPacket filtering"
-  command = "sudo sh -c \"pfctl -e\""
-  execute_command(message,command)
-  message = "Loading:\yFilters from "+pf_file
-  command = "sudo sh -c \"pfctl -F all -f #{pf_file}\""
-  execute_command(message,command)
   return
 end
